Improve organization of the Compose code
